change WorkerPool and Worker signature
